Support for 64bit and Xcode 5.1 warning suppressions

+++ Classes/Mod.m
@@ -45,7 +45,7 @@
     
 	if (modCategory) {
         // fire request to moderate the post
-        NSString *modUrl = [NSString stringWithFormat:@"http://www.shacknews.com/mod_chatty.x?root=%d&post_id=%d&mod_type_id=%@", parentId, postId, modCategory];
+        NSString *modUrl = [NSString stringWithFormat:@"http://www.shacknews.com/mod_chatty.x?root=%lu&post_id=%lu&mod_type_id=%@", (unsigned long)parentId, (unsigned long)postId, modCategory];
 		NSMutableURLRequest *modRequest;
         //NSLog(@"Moderating post with URL: %@", url);
 		modRequest = [[NSMutableURLRequest alloc] initWithURL:[NSURL URLWithString:modUrl]
@@ -61,7 +61,7 @@
                                                       cachePolicy:NSURLRequestReloadIgnoringCacheData
                                                   timeoutInterval:60];
         [reindexRequest setHTTPMethod:@"POST"];
-        [reindexRequest setHTTPBody:[[NSString stringWithFormat:@"postId=%i", postId] dataUsingEncoding:NSASCIIStringEncoding]];
+        [reindexRequest setHTTPBody:[[NSString stringWithFormat:@"postId=%lu", (unsigned long)postId] dataUsingEncoding:NSASCIIStringEncoding]];
         [NSURLConnection connectionWithRequest:reindexRequest delegate:nil];
 
         // test response bodies
