+++ Classes/Tag.m
@@ -12,7 +12,7 @@
 
 + (void)tagPostId:(NSUInteger)postId tag:(NSString*)tag {
     NSString *username = [[NSUserDefaults standardUserDefaults] stringForKey:@"username"];
-    NSString *url = [[NSString stringWithFormat:@"http://lmnopc.com/greasemonkey/shacklol/report.php?who=%@&what=%d&tag=%@&version=-1", username, postId, tag] stringByAddingPercentEscapesUsingEncoding:NSASCIIStringEncoding];
+    NSString *url = [[NSString stringWithFormat:@"http://lmnopc.com/greasemonkey/shacklol/report.php?who=%@&what=%lu&tag=%@&version=-1", username, (unsigned long)postId, tag] stringByAddingPercentEscapesUsingEncoding:NSASCIIStringEncoding];
     
     NSLog(@"Tagging post with URL: %@", url);
     NSURLRequest *request = [NSURLRequest requestWithURL:[NSURL URLWithString:url]];
